requirements.txt
argcomplete==2.0.0
asgiref==3.6.0
boto3==1.26.59
botocore==1.29.59
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.4
django-storages==1.13.2
durationpy==0.5
Faker==16.6.1
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.2
mysqlclient==2.1.1
placebo==0.9.0
psycopg2-binary==2.9.5
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
toposort==1.9
tqdm==4.64.1
troposphere==4.3.0
urllib3==1.26.14
Werkzeug==2.2.2
zappa==0.56.1
